requirements.txt
cachetools==5.4.0 ; python_version >= "3.8" and python_version < "3.11"
certifi==2024.7.4 ; python_version >= "3.8" and python_version < "3.11"
charset-normalizer==3.3.2 ; python_version >= "3.8" and python_version < "3.11"
contourpy==1.1.1 ; python_version >= "3.8" and python_version < "3.11"
cycler==0.12.1 ; python_version >= "3.8" and python_version < "3.11"
db-dtypes==1.2.0 ; python_version >= "3.8" and python_version < "3.11"
fonttools==4.53.1 ; python_version >= "3.8" and python_version < "3.11"
google-api-core==2.19.1 ; python_version >= "3.8" and python_version < "3.11"
google-api-core[grpc]==2.19.1 ; python_version >= "3.8" and python_version < "3.11"
google-auth==2.32.0 ; python_version >= "3.8" and python_version < "3.11"
google-cloud-bigquery-storage==2.25.0 ; python_version >= "3.8" and python_version < "3.11"
google-cloud-bigquery==3.25.0 ; python_version >= "3.8" and python_version < "3.11"
google-cloud-core==2.4.1 ; python_version >= "3.8" and python_version < "3.11"
google-crc32c==1.5.0 ; python_version >= "3.8" and python_version < "3.11"
google-resumable-media==2.7.1 ; python_version >= "3.8" and python_version < "3.11"
googleapis-common-protos==1.63.2 ; python_version >= "3.8" and python_version < "3.11"
grpcio-status==1.62.3 ; python_version >= "3.8" and python_version < "3.11"
grpcio==1.65.4 ; python_version >= "3.8" and python_version < "3.11"
idna==3.7 ; python_version >= "3.8" and python_version < "3.11"
kiwisolver==1.4.5 ; python_version >= "3.8" and python_version < "3.11"
markdown-it-py==3.0.0 ; python_version >= "3.8" and python_version < "3.11"
matplotlib==3.6.0 ; python_version >= "3.8" and python_version < "3.11"
mdurl==0.1.2 ; python_version >= "3.8" and python_version < "3.11"
numpy==1.24.4 ; python_version >= "3.8" and python_version < "3.11"
packaging==24.1 ; python_version >= "3.8" and python_version < "3.11"
pandas==2.0.3 ; python_version >= "3.8" and python_version < "3.11"
pillow==10.4.0 ; python_version >= "3.8" and python_version < "3.11"
plotly==5.23.0 ; python_version >= "3.8" and python_version < "3.11"
proto-plus==1.24.0 ; python_version >= "3.8" and python_version < "3.11"
protobuf==4.25.4 ; python_version >= "3.8" and python_version < "3.11"
pyarrow==17.0.0 ; python_version >= "3.8" and python_version < "3.11"
pyasn1-modules==0.4.0 ; python_version >= "3.8" and python_version < "3.11"
pyasn1==0.6.0 ; python_version >= "3.8" and python_version < "3.11"
pygments==2.18.0 ; python_version >= "3.8" and python_version < "3.11"
pyparsing==3.1.2 ; python_version >= "3.8" and python_version < "3.11"
python-dateutil==2.9.0.post0 ; python_version >= "3.8" and python_version < "3.11"
pytz==2024.1 ; python_version >= "3.8" and python_version < "3.11"
requests==2.32.3 ; python_version >= "3.8" and python_version < "3.11"
rich==13.7.1 ; python_version >= "3.8" and python_version < "3.11"
rsa==4.9 ; python_version >= "3.8" and python_version < "3.11"
scipy==1.10.1 ; python_version >= "3.8" and python_version < "3.11"
seaborn==0.13.2 ; python_version >= "3.8" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.8" and python_version < "3.11"
tenacity==9.0.0 ; python_version >= "3.8" and python_version < "3.11"
typing-extensions==4.12.2 ; python_version >= "3.8" and python_version < "3.9"
tzdata==2024.1 ; python_version >= "3.8" and python_version < "3.11"
urllib3==2.2.2 ; python_version >= "3.8" and python_version < "3.11"
